feat: Implement chat list screen and update ChatScreen for dynamic chat loading
- Created ChatsListScreen to fetch and display user chats from the API.
- Each chat is rendered as a card with the last message.
- Added navigation to ChatScreen with the selected chat ID.
- Updated ChatScreen to accept chatId as a parameter for loading messages.
- Integrated Expo Router for navigation handling.
